Move Keypairs into its own module
