+++ client/src/thin_client.rs
@@ -14,13 +14,14 @@ use solana_sdk::{
     fee_calculator::FeeCalculator,
     hash::Hash,
     instruction::Instruction,
+    keypairs::Keypairs,
     message::Message,
     packet::PACKET_DATA_SIZE,
     pubkey::Pubkey,
     signature::{Keypair, KeypairUtil, Signature},
     system_instruction,
     timing::duration_as_ms,
-    transaction::{self, Keypairs, Transaction},
+    transaction::{self, Transaction},
     transport::Result as TransportResult,
 };
 use std::{
